requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
gremlinpython==3.7.0
websockets==12.0
httpx==0.25.2
sentence-transformers==2.3.1
huggingface_hub==0.19.4
transformers==4.36.0
faiss-cpu==1.7.4
google-generativeai==0.3.0
numpy==1.24.3
scikit-learn==1.3.0
tenacity==8.2.3
loguru==0.7.2
prometheus-client==0.19.0
psutil==5.9.6
torch>=2.0.0
requests
